requirements.txt
cryptography==3.4.8
PyYAML==5.4.1
paramiko==2.7.2